+++ spec/frontend/support/commands.js
@@ -1,25 +1,23 @@
-// ***********************************************
-// This example commands.js shows you how to
-// create various custom commands and overwrite
-// existing commands.
-//
-// For more comprehensive examples of custom
-// commands please read more here:
-// https://on.cypress.io/custom-commands
-// ***********************************************
-//
-//
-// -- This is a parent command --
-// Cypress.Commands.add("login", (email, password) => { ... })
-//
-//
-// -- This is a child command --
-// Cypress.Commands.add("drag", { prevSubject: 'element'}, (subject, options) => { ... })
-//
-//
-// -- This is a dual command --
-// Cypress.Commands.add("dismiss", { prevSubject: 'optional'}, (subject, options) => { ... })
-//
-//
-// -- This is will overwrite an existing command --
-// Cypress.Commands.overwrite("visit", (originalFn, url, options) => { ... })
+Cypress.Commands.add('login', (email = 'email@example.com', password = 'password2018') => {
+  return cy.request({
+    method: 'POST',
+    url: '/admin/login',
+    body: {
+      user: {
+        email: email,
+        password: password
+      }
+    }
+  })
+})
+
+Cypress.Commands.add('logout', () => {
+  return cy.request({
+    method: 'DELETE',
+    url: '/admin/logout',
+  })
+})
+
+Cypress.Commands.add('setSessionCookies', () => {
+  Cypress.Cookies.preserveOnce('_gmmcal_session')
+})